interim

+++ BoardForth.ino
@@ -1,3 +1,4 @@
+
 #include "vm.h"
 
 // FLASH(ok) = ": ok cr 'o' emit 'k' emit cr ; ok";
@@ -87,7 +88,7 @@ void setup() {
     parseLine(buf);
   }
   writePort_String("\nBoardForth v0.0.1 Written by Chris Curl (2021)");
-  writePort_StringF("\nDictionary size is: %d (0x%lx) bytes.", DICT_SZ, DICT_SZ);
+  writePort_StringF("\nDictionary size is: %d (0x%lx) bytes.", DICT_SZ, (long)DICT_SZ);
   writePort_String("\nHello.");
 }
